+++ memory_bus.h
@@ -7,8 +7,8 @@
 
 typedef struct
 {
-	int offset;
-	int mask;
+	uint64_t offset;
+	uint64_t mask;
 	memory *target;
 } memory_segment_t;
 
@@ -17,21 +17,21 @@ class memory_bus
 private:
 	std::vector<memory_segment_t> list;
 
-	const memory_segment_t * find_segment(int offset) const;
+	const memory_segment_t * find_segment(uint64_t offset) const;
 
 public:
 	memory_bus();
 	~memory_bus();
 
-	void register_memory(int offset, int mask, memory *target);
+	void register_memory(uint64_t offset, uint64_t mask, memory *target);
 
-	bool read_64b(int offset, uint64_t *data) const;
-	bool write_64b(int offset, uint64_t data);
-	bool read_32b(int offset, uint32_t *data) const;
-	bool write_32b(int offset, uint32_t data);
-	bool read_16b(int offset, uint16_t *data) const;
-	bool write_16b(int offset, uint16_t data);
-	bool read_8b(int offset, uint8_t *data) const;
-	bool write_8b(int offset, uint8_t data);
+	bool read_64b(uint64_t offset, uint64_t *data) const;
+	bool write_64b(uint64_t offset, uint64_t data);
+	bool read_32b(uint64_t offset, uint32_t *data) const;
+	bool write_32b(uint64_t offset, uint32_t data);
+	bool read_16b(uint64_t offset, uint16_t *data) const;
+	bool write_16b(uint64_t offset, uint16_t data);
+	bool read_8b(uint64_t offset, uint8_t *data) const;
+	bool write_8b(uint64_t offset, uint8_t data);
 };
 #endif
